Switch from z.lua to Zoxide
Switches from z.lua to Zoxide and sets up the ability to automatically
compile and install Rust programs to be used by zsh plugins.

This will result in a relatively long compile on the initial run of ZSH
but will setup the tools required for ZSH to work with it's plugins.

To import the z database from z.lua run:

```
zoxide import --from z ~/.zlua --merge
```