Remove lspinstall
Removes lspinstall as I have realised that managing the lsp's using this
method is probably silly for the following reason... updates... it is
quite likely that we will forget to update the lsp's frequently so it is
a better idea to install these outside of Neovim using scripts / Ansible
or other configuration management and package managers so that the
package managers are managing them directly as part of the OS.
Alternatively we could have the lsps run inside Docker.

I have refactored the lspconfig configuration to use whatever version of
the lsps are installed and available on the PATH. Only terraformls, lua
and yamlls have been implemented so far.

Align plugin comments

Initial testing of lspconfig iterator

Current progress

Initial LSP refactor

Refactored